pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
exclude = ["resources/*", "tests/*", "dataset/*"]

[tool.hatch.build.targets.sdist]
exclude = ["resources/*", "tests/*", "dataset/*"]

[project]
name = "PopV"
version = "0.4.2"
description = "Consensus prediction of cell type labels with popV"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Can Ergen"},
]
maintainers = [
    {name = "Can Ergen", email = "canergen.ac@gmail.com"},
]
urls.Documentation = "https://PopV.readthedocs.io/"
urls.Source = "https://github.com/YosefLab/PopV.git"
urls.Home-page = "https://github.com/YosefLab/PopV.git"

dependencies = [
    "anndata>0.8.0",
    "pandas<2.0.0,>=1.4.0",
    "bbknn>1.5.0",
    "celltypist>1.3.0",
    "gdown>4.6.0",
    "h5py>3.7.0",
    "huggingface-hub>=0.11.1",
    "onclass>=1.3",
    "scanorama>1.7.2",
    "scanpy>1.9.0",
    "scikit-learn>0.21.2,<1.2",
    "scikit-misc>=0.1",
    "scvi-tools>1.0.0",
    "tensorflow>2.11.0",
    "transformers>4.25.0",
    "obonet>=1.0",
    "harmony-pytorch>0.1.6",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "twine>=4.0.2"
]
doc = [
    "docutils>=0.8,!=0.18.*,!=0.19.*",
    "sphinx>=4",
    "sphinx-book-theme>=1.0.0",
    "myst-nb",
    "sphinxcontrib-bibtex>=1.0.0",
    "sphinx-autodoc-typehints",
    "sphinxext-opengraph",
    # For notebooks
    "ipykernel",
    "ipython",
    "sphinx-copybutton",
    "pandas",
]
test = [
    "pytest",
    "coverage",
]

[tool.coverage.run]
source = ["popv"]
omit = [
    "**/test_*.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",  # allow using test files with same name
]

[tool.black]
line-length = 120

[tool.ruff]
src = ["."]
line-length = 119
target-version = "py39"
lint.select = [
    "F",  # Errors detected by Pyflakes
    "E",  # Error detected by Pycodestyle
    "W",  # Warning detected by Pycodestyle
    "I",  # isort
    "D",  # pydocstyle
    "B",  # flake8-bugbear
    "TID",  # flake8-tidy-imports
    "C4",  # flake8-comprehensions
    "BLE",  # flake8-blind-except
    "UP",  # pyupgrade
    "RUF",  # Ruff formatting
    "TCH", # flake8-type-checking
    "FA", # flake8-future-annotations (FA)
    "TD", # TODO
    "UP",  # pyupgrade
    "RUF",  # Report unused noqa directives
    "PT", # pytest style
]
lint.ignore = [
    # line too long -> we accept long comment lines; black gets rid of long code lines
    "E501",
    # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
    "E731",
    # allow I, O, l as variable names -> I is the identity matrix
    "E741",
    # Missing docstring in public package
    "D104",
    # Missing docstring in public module
    "D100",
    # Missing docstring in __init__
    "D107",
    # Errors from function calls in argument defaults. These are fine when the result is immutable.
    "B008",
    # __magic__ methods are are often self-explanatory, allow missing docstrings
    "D105",
    # first line should end with a period [Bug: doesn't work with single-line docstrings]
    "D400",
    # First line should be in imperative mood; try rephrasing
    "D401",
    ## Disable one in each pair of mutually incompatible rules
    # We don’t want a blank line before a class docstring
    "D203",
    # We want docstrings to start immediately after the opening triple quote
    "D213",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"docs/*" = ["I", "BLE001"]
"tests/*" = ["D"]
"*/__init__.py" = ["F401"]

[tool.cruft]
skip = [
    "tests",
    "src/**/__init__.py",
    "src/**/basic.py",
    "docs/api.md",
    "docs/changelog.md",
    "docs/references.bib",
    "docs/references.md",
    "docs/notebooks/example.ipynb"
]
